+++ lib/src/iso/bmff/filterBoxes.ts
@@ -1,4 +1,3 @@
-import type { Box } from './boxes/Box.js';
 import type { ContainerBox } from './boxes/ContainerBox.js';
 import type { IsoBmffBox } from './boxes/IsoBmffBox.js';
 import type { BoxFilter } from './BoxFilter.js';
@@ -6,7 +5,7 @@ import { createIsoView } from './createIsoView.js';
 import type { IsoData } from './IsoData.js';
 import type { IsoViewConfig } from './IsoViewConfig.js';
 
-function filter<T extends Box = IsoBmffBox>(iterator: Iterable<IsoBmffBox>, fn: BoxFilter, recursive: boolean, result: T[]): T[] {
+function filter<T extends IsoBmffBox = IsoBmffBox>(iterator: Iterable<IsoBmffBox>, fn: BoxFilter<T>, recursive: boolean, result: T[]): T[] {
 	for (const box of iterator) {
 		if (fn(box)) {
 			result.push(box as T);
@@ -32,11 +31,11 @@ function filter<T extends Box = IsoBmffBox>(iterator: Iterable<IsoBmffBox>, fn: 
  * @group ISOBMFF
  * @beta
  */
-export function filterBoxes<T extends Box = IsoBmffBox>(raw: IsoData | Iterable<IsoBmffBox>, fn: BoxFilter, config?: IsoViewConfig): T[] {
+export function filterBoxes<T extends IsoBmffBox = IsoBmffBox>(raw: IsoData | Iterable<IsoBmffBox>, fn: BoxFilter<T>, config?: IsoViewConfig): T[] {
 	if (raw instanceof DataView || raw instanceof Uint8Array || raw instanceof ArrayBuffer) {
 		raw = createIsoView(raw, { ...config, recursive: false });
 	}
 
 	const recursive = config?.recursive ?? true;
-	return filter(raw, fn, recursive, []);
+	return filter<T>(raw, fn, recursive, []);
 }