+++ lib/src/iso/bmff/findBoxByType.ts
@@ -1,5 +1,5 @@
-import type { Box } from './boxes/Box.js';
 import type { IsoBmffBox } from './boxes/IsoBmffBox.js';
+import type { IsoBmffBoxMap } from './boxes/IsoBmffBoxMap.js';
 import { findBox } from './findBox.js';
 import type { IsoData } from './IsoData.js';
 import type { IsoViewConfig } from './IsoViewConfig.js';
@@ -17,10 +17,6 @@ import type { IsoViewConfig } from './IsoViewConfig.js';
  *
  * @beta
  */
-export function findBoxByType<T extends Box = IsoBmffBox>(raw: IsoData, type: string, config: IsoViewConfig = {}): T | null {
-	return findBox<T>(raw, box => box.type === type, config);
+export function findBoxByType<T extends keyof IsoBmffBoxMap>(raw: IsoData, type: T, config: IsoViewConfig = {}): IsoBmffBoxMap[T] | null {
+	return findBox(raw, (box: IsoBmffBox): box is IsoBmffBoxMap[T] => box.type === type, config);
 }
-
-const box = findBoxByType(new ArrayBuffer(100), 'ftyp', { recursive: true });
-
-console.log(box);